+++ src/main/kotlin/com/pestphp/pest/PestIcons.kt
@@ -1,8 +1,10 @@
 package com.pestphp.pest
 
+import com.intellij.icons.AllIcons
 import com.intellij.ide.ui.LafManager
 import com.intellij.ide.ui.laf.darcula.DarculaLookAndFeelInfo
 import com.intellij.openapi.util.IconLoader
+import com.jetbrains.php.PhpIcons
 
 object PestIcons {
     val LOGO = IconLoader.getIcon("/logo@${getThemeString()}.svg", PestIcons.javaClass)
@@ -10,6 +12,7 @@ object PestIcons {
     val CONFIG = IconLoader.getIcon("/config@${getThemeString()}.svg", PestIcons.javaClass)
     val FILE = IconLoader.getIcon("/file@${getThemeString()}.svg", PestIcons.javaClass)
     val DATASET_FILE = IconLoader.getIcon("/dataset@${getThemeString()}.svg", PestIcons.javaClass)
+    val SNAPSHOT_ICON = AllIcons.Nodes.DataSchema
 
     private fun isDarkMode() = LafManager.getInstance().currentLookAndFeel == DarculaLookAndFeelInfo()
     private fun getThemeString(): String {