+++ src/test/kotlin/com/pestphp/pest/customExpectations/generators/ExpectationGeneratorTest.kt
@@ -1,14 +1,13 @@
-package com.pestphp.pest.generators
+package com.pestphp.pest.customExpectations.generators
 
 import com.jetbrains.php.lang.psi.resolve.types.PhpType
 import com.pestphp.pest.PestLightCodeFixture
-import junit.framework.TestCase
 
 class ExpectationGeneratorTest : PestLightCodeFixture() {
     fun testCanGenerateFileWithMethod() {
         val generator = ExpectationGenerator()
         generator.docMethods.add(
-            ExpectationGenerator.Method(
+            Method(
                 "works",
                 PhpType.BOOLEAN,
                 emptyList()
@@ -18,7 +17,7 @@ class ExpectationGeneratorTest : PestLightCodeFixture() {
         val result = generator.generate(project)
 
         val expectedClass = ExpectationGeneratorTest::class.java
-            .getResource("/com/pestphp/pest/generators/ExpectationGenerator/GeneratedWithMethod.php")
+            .getResource("/com/pestphp/pest/customExpectations/generators/ExpectationGenerator/GeneratedWithMethod.php")
             ?.readText() ?: fail("File not found.")
         assertEquals(
             result,