+++ src/test/resources/com/pestphp/pest/customExpectations/CustomExpectation.php
@@ -1,0 +1,5 @@
+<?php
+
+expect()->extend('toBeOne', function (): bool {
+    return true;
+});