+++ src/test/resources/com/pestphp/pest/customExpectations/CustomThisExpectation.php
@@ -1,0 +1,5 @@
+<?php
+
+expect()->extend('toThis', function () {
+    return $this->toBeEmpty();
+});