+++ src/test/resources/com/pestphp/pest/customExpectations/CustomUserExpectation.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Models;
+
+class User {
+
+}
+
+expect()->extend('createUser', function (): User {
+    return new User();
+});