+++ src/test/resources/com/pestphp/pest/customExpectations/subFolder/CustomExpectation.php
@@ -1,0 +1,5 @@
+<?php
+
+expect()->extend('toBeExpectationFromSubFolder', function (): bool {
+    return true;
+});