+++ src/test/resources/com/pestphp/pest/snapshotTesting/allSnapshotAssertions.php
@@ -1,0 +1,11 @@
+<?php
+
+assertMatchesSnapshot("test");
+assertMatchesFileHashSnapshot("test");
+assertMatchesFileSnapshot("test");
+assertMatchesHtmlSnapshot("test");
+assertMatchesJsonSnapshot("test");
+assertMatchesObjectSnapshot("test");
+assertMatchesTextSnapshot("test");
+assertMatchesXmlSnapshot("test");
+assertMatchesYamlSnapshot("test");